requirements.txt
pandas
scikit-learn
streamlit
scikit-surprise
numpy<2.0